--- src/transform.rs
@@ -1,18 +1,0 @@
-mod crs;
-
-// pub struct Transform<Dest> {
-//     prev: Transform<Dest>,
-// }
-// impl Transform<()> {
-//     fn new() -> Self {
-//         Self { inner: () }
-//     }
-// }
-// impl Transform<crs::Tokyo> {
-//     pub fn from_tokyo() -> Self {
-//         Self { inner: crs::Tokyo }
-//     }
-//     pub fn to_jgd2000() -> Transform<Transform<crs::JGD2000>> {
-//         todo!()
-//     }
-// }
